+++ CHANGELOG.md
@@ -4,8 +4,7 @@
 
 * PayPalNativeCheckout
   * Bump native-checkout version to release `1.2.0`
-  * Adding in support for the merchant passed email feature (opt in only)
-
+  * add `setUserAuthenticationEmail() to PayPalNativeRequest for merchant passed email feature (opt in only)
 * GooglePay
   * Bump `play-services-wallet` version to `19.2.1`
   * Add `totalPriceLabel` to `GooglePayRequest`